# services/chat/api/openapi.yaml
openapi: 3.0.4
info:
  title: Chat Service API
  version: 1.0.0
  description: API for managing messages and friendships
tags:
  - name: Friendship
    description: Endpoints for managing friendships
  - name: Message
    description: Endpoints for managing chat messages

components:
  schemas:
    ChatRoom:
      type: object
      required: [chatId, type]
      properties:
        chatId:
          type: string
          format: uuid
        type:
          type: string
          enum: [direct, group]
        name:
          type: string
          nullable: true
          description: "Will be null for direct chats"
    ChatMessage:
      type: object
      required: [id, chatId, senderId, content, createdAt]
      properties:
        id:
          type: integer
        chatId:
          type: string
          format: uuid
        senderId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time
          example: "2026-03-09T08:24:27Z"
    FriendList:
      type: array
      items:
        type: object
        properties:
          friendId:
            type: integer
            description: Unique identifier for the friend
          chatId:
            type: string
            format: uuid
            description: Unique identifier for the chat between the user and the friend
      example:
        friendId: 123
        chatId: "550e8400-e29b-41d4-a716"
    StreamEvent:
      type: object
      required: [type, payload]
      properties:
        type:
          type: string
          enum: [NEW_MESSAGE, USER_TYPING, USER_STATUS]
          description: "The type of event being sent down the stream"
        payload:
          type: object
          description: "The actual data, structure depends on the event type"
          # We use oneOf to say "If it's a message, expect ChatMessage data"
          oneOf:
            - $ref: "#/components/schemas/ChatMessage"
            # You can add other schemas here later, like TypingIndicator!

paths:
  /friendship/{requesterId}/{receiverId}:
    post:
      tags:
        - Friendship
      summary: Send a friend request
      operationId: sendFriendRequest
      parameters:
        - name: requesterId
          in: path
          required: true
          schema:
            type: integer
        - name: receiverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Friend request sent successfully
        "400":
          description: Invalid user IDs
        "404":
          description: User not found

  /friendship/{requesterId}/{receiverId}/update:
    patch:
      tags:
        - Friendship
      summary: Update friendship status
      operationId: updateFriendshipStatus
      parameters:
        - name: requesterId
          in: path
          required: true
          schema:
            type: integer
        - name: receiverId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          required: true
          schema:
            type: string
            enum: [pending, accepted, declined, blocked, none]
      responses:
        "200":
          description: Status updated successfully
        "400":
          description: Invalid status value
        "404":
          description: Request not found
        "500":
          description: Internal Server Error

  /friendship/{tempUserId}:
    get:
      tags:
        - Friendship
      summary: Get friend list for a user with chat ID references
      operationId: getFriendList
      parameters:
        - name: tempUserId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Friend list and associated chat IDs retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendList"
        "400":
          description: Invalid user ID
        "404":
          description: User not found
        "500":
          description: Internal Server Error

  /message/{chatId}/{tempSenderId}:
    post:
      tags:
        - Message
      summary: Send a chat message
      operationId: sendMessage
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: tempSenderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 1
      responses:
        "201":
          description: Message sent successfully
        "400":
          description: Invalid user IDs or message content

  /message/history/{chatId}:
    get:
      tags:
        - Message
      summary: Get chat history for a specific chat
      operationId: getMessageHistory
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Chat history retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
        "400":
          description: Invalid chat ID
        "404":
          description: No chat history found

  /message/stream/{tempUserId}:
    get:
      tags:
        - Message
      summary: Opens an SSE connection for real-time chat message updates
      operationId: getMessageStream
      parameters:
        - name: tempUserId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SSE connection successful
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/StreamEvent"
        "401":
          description: Unauthorized

  # /chats/{tempUserId}:
  #   get:
  #     tags:
  #       - Chat
  #     summary: Get all chats (direct and group) for a user's inbox
  #     operationId: getUserChats
  #     parameters:
  #       - name: tempUserId
  #         in: path
  #         required: true
  #         schema:
  #           type: integer
  #     responses:
  #       "200":
  #         description: Successfully retrieved user's chat rooms
  #         content:
  #           application/json:
  #             schema:
  #               type: array
  #               items:
  #                 $ref: "#/components/schemas/ChatRoom"

  # /chats/group:
  #   post:
  #     tags:
  #       - Chat
  #     summary: Create a new group chat
  #     operationId: createGroupChat
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             type: object
  #             required: [name, memberIds]
  #             properties:
  #               name:
  #                 type: string
  #                 description: "Name of the group chat"
  #               memberIds:
  #                 type: array
  #                 items:
  #                   type: integer
  #                 description: "List of user IDs to add to the group"
  #     responses:
  #       "201":
  #         description: Group chat created successfully
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ChatRoom"